backend/rag_bot/requirements.txt
cohere>=5.0.0
qdrant-client>=1.8.0
python-dotenv>=1.0.0